Fix bugs in the codebase due to Travis CI not being functional for continuous integration

+++ server/src/main/java/com/github/jhu_oose11/calendue/controllers/AssignmentsController.java
@@ -26,8 +26,8 @@ public class AssignmentsController {
         String title = ctx.formParam("title");
         LocalDate dueDate = LocalDate.parse(Objects.requireNonNull(ctx.formParam("due_date")));
         String gradeString = ctx.formParam("grade");
-        double score = Integer.parseInt(gradeString.split("/")[0]);
-        double total = Integer.parseInt(gradeString.split("/")[1]);
+        double score = Double.parseDouble(gradeString.split("/")[0]);
+        double total = Double.parseDouble(gradeString.split("/")[1]);
         double grade = score/total * 100;
 
         int course_id;
